adpredictor-v2/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.19
pydantic[email-validator]==2.10.3
python-dotenv==1.0.1
stripe==11.3.0
anthropic==0.39.0
ffmpeg-python==0.2.0
pillow==11.1.0